requirements.txt
quixstreams>=2.4.0
kafka-python
pandas
matplotlib